Bring back Byte32De
